+++ man-roxygen/params-netlogo-home.R
@@ -1,5 +1,5 @@
 #' @param netlogo_home (optional) A [`character`][base::character] string
 #'   specifying the path to the NetLogo installation directory. If not provided,
 #'   the function will try to find it automatically using
-#'   [`find_netlogo_home()`][find_netlogo_home]. (default:
-#'   [`find_netlogo_home()`][find_netlogo_home]).
+#'   [`find_netlogo_home()`][find_netlogo_home]
+#'   (default: `find_netlogo_home()`).